corrige le calcul d’un statut d’une démarche
