+++ packages/api/src/business/processes/titres-etapes-ordre-update.ts
@@ -2,7 +2,7 @@ import { titreEtapeUpdate } from '../../database/queries/titres-etapes.js'
 import { titreEtapesSortAscByDate } from '../utils/titre-etapes-sort.js'
 import { DemarcheTypeId } from 'camino-common/src/static/demarchesTypes.js'
 import { TitreTypeId } from 'camino-common/src/static/titresTypes.js'
-import { getDemarches } from './titres-etapes-heritage-contenu-update.js'
+import { getDemarches } from './titres-etapes-heritage-contenu-update.queries.js'
 import { UserNotNull } from 'camino-common/src/roles.js'
 import { DemarcheId } from 'camino-common/src/demarche.js'
 import { Pool } from 'pg'
